splitting up config types into their respective modules
